Updated layout.
