+++ src/UI/Balance/balanceStyles.scss
@@ -31,6 +31,7 @@
       font-size: 66%;
       position: relative;
       top: -0.125em;
+      vertical-align: unset;
     }
   }
 }
