+++ src/UI/SignTransactionsModals/SignWithDeviceModal/components/components/ConfirmAmount/components/ConfirmAmountData/index.ts
@@ -1,0 +1,1 @@
+export * from './ConfirmAmountData';
