+++ src/UI/SignTransactionsModals/SignWithDeviceModal/components/components/ConfirmAmount/components/index.ts
@@ -1,1 +1,2 @@
 export * from './ConfirmAmountLabel';
+export * from './ConfirmAmountData';
