+++ src/UI/SignTransactionsModals/SignWithDeviceModal/components/components/ConfirmAmount/confirmAmountStyles.scss
@@ -1,33 +1,96 @@
-.amount {
+.confirmAmount {
   display: flex;
   flex-direction: column;
-  text-align: left;
+  line-height: 1;
+  gap: 8px;
 
-  .label {
-    margin-bottom: 8px;
-    color: #a3a3a3;
-    font-size: 12px;
-    line-height: 1;
-    font-weight: 400;
-    display: block;
-  }
+  .confirmAmountWrapper {
+    height: 48px;
+    position: relative;
 
-  .token {
-    display: flex;
-    align-items: center;
+    .confirmAmountLoading {
+      opacity: 0;
+      pointer-events: none;
+      position: absolute;
+      left: 0;
+      top: 0;
+      right: 0;
+      bottom: 0;
 
-    .value {
-      font-size: 14px;
-      font-weight: 500;
-      color: #e5e5e5;
-      line-height: 1.5;
+      &.confirmAmountLoadingVisible {
+        opacity: 1;
+      }
     }
-  }
 
-  .price {
-    font-size: 12px;
-    color: #737373;
-    line-height: 1;
-    margin-top: 8px;
+    .confirmAmountContent {
+      position: absolute;
+      left: 0;
+      top: 0;
+      right: 0;
+      bottom: 0;
+
+      &.confirmAmountContentHidden {
+        pointer-events: none;
+        opacity: 0;
+      }
+
+      .confirmAmountPreview {
+        display: flex;
+        cursor: pointer;
+        justify-content: space-between;
+        align-items: center;
+
+        &:hover .confirmAmountPreviewIcon {
+          color: #0ac2ae;
+        }
+
+        .confirmAmountPreviewWrapper {
+          gap: 8px;
+          display: flex;
+          align-items: center;
+
+          .confirmAmountPreviewIconWrapper {
+            position: relative;
+
+            .confirmAmountPreviewIcon {
+              position: absolute;
+              top: 0;
+              left: 0;
+              width: 48px;
+              overflow: hidden;
+              border-radius: 8px;
+              height: 48px;
+              border: 1px solid #2e2e2e;
+
+              &.confirmAmountPreviewIconRelative {
+                position: relative;
+              }
+            }
+          }
+
+          .confirmAmountPreviewContent {
+            display: flex;
+            flex-direction: column;
+            gap: 4px;
+
+            .confirmAmountPreviewName {
+              font-size: 16px;
+              color: #e5e5e5;
+              line-height: 1;
+            }
+
+            .confirmAmountPreviewTicker {
+              color: #525252;
+              font-size: 10px;
+            }
+          }
+        }
+
+        .confirmAmountPreviewIcon {
+          color: #525252;
+          transition: all 200ms ease-out;
+        }
+      }
+    }
   }
 }
