--- src/UI/SignTransactionsModals/SignWithDeviceModal/components/components/ConfirmAmount2/index.ts
@@ -1,1 +1,0 @@
-export * from './ConfirmAmount';
