+++ src/UI/SignTransactionsModals/SignWithDeviceModal/components/components/ConfirmFee/confirmFeeStyles.scss
@@ -11,7 +11,7 @@
   .confirmFeeData {
     display: flex;
     align-items: center;
-    gap: 12px;
+    gap: 8px;
 
     .confirmFeeDataPriceWrapper {
       display: flex;
