--- src/UI/SignTransactionsModals/SignWithDeviceModal/components/components/NftSftPreviewComponent/NftSftPreviewComponent.styles.scss
@@ -1,93 +1,0 @@
-.preview {
-  border-radius: 8px;
-  display: flex;
-  align-items: center;
-  cursor: pointer;
-  overflow: hidden;
-  background-color: #262626;
-
-  .image {
-    min-width: 64px;
-    min-height: 64px;
-    max-width: 64px;
-    max-height: 64px;
-    height: auto;
-    width: auto;
-  }
-
-  .content {
-    padding: 12px;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    width: 100%;
-
-    .left {
-      gap: 8px;
-      line-height: 1;
-      display: flex;
-      flex-direction: column;
-
-      .name {
-        font-size: 12px;
-        font-weight: 600;
-        text-align: left;
-        color: #e5e5e5;
-      }
-
-      .identifier {
-        color: #737373;
-        font-size: 12px;
-      }
-    }
-
-    .badge {
-      line-height: 1;
-      font-weight: 500;
-      font-size: 12px;
-      padding: 4px 8px;
-      text-transform: uppercase;
-      border-radius: 4px;
-      background-color: #171717;
-      position: relative;
-
-      &:before {
-        content: '';
-        border-radius: 4px;
-        position: absolute;
-        inset: 0;
-        padding: 1px;
-        mask: linear-gradient(#fff 0 0) content-box, linear-gradient(#fff 0 0);
-        -webkit-mask-composite: xor;
-        -webkit-mask: linear-gradient(#fff 0 0) content-box,
-          linear-gradient(#fff 0 0);
-        mask-composite: exclude;
-        pointer-events: none;
-      }
-
-      &.nft {
-        color: #fef08a;
-
-        &:before {
-          background: linear-gradient(
-            90deg,
-            rgba(240, 152, 25, 0.4),
-            rgba(237, 222, 93, 0.4)
-          );
-        }
-      }
-
-      &.sft {
-        color: #fed7aa;
-
-        &:before {
-          background: linear-gradient(
-            90deg,
-            rgba(240, 77, 25, 0.4),
-            rgba(237, 136, 93, 0.4)
-          );
-        }
-      }
-    }
-  }
-}
