--- src/UI/SignTransactionsModals/SignWithDeviceModal/components/components/NftSftPreviewComponent/index.ts
@@ -1,1 +1,0 @@
-export * from './NftSftPreviewComponent';
