--- src/UI/SignTransactionsModals/SignWithDeviceModal/components/components/TokenAvatar/index.ts
@@ -1,1 +1,0 @@
-export * from './TokenAvatar';
