--- src/UI/SignTransactionsModals/SignWithDeviceModal/components/components/TokenAvatar/tokenAvatarStyles.scss
@@ -1,61 +1,0 @@
-.tokenAvatar {
-  width: 20px;
-  height: 20px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  line-height: 1;
-  text-transform: uppercase;
-  border-radius: 50%;
-  min-width: 20px;
-  min-height: 20px;
-  border: 1px solid #737373;
-  margin-right: 4px;
-  position: relative;
-  font-size: 6px;
-
-  &:before {
-    content: '';
-    border-radius: 50%;
-    position: absolute;
-    inset: 0;
-    padding: 1px;
-    mask: linear-gradient(#fff 0 0) content-box, linear-gradient(#fff 0 0);
-    -webkit-mask-composite: xor;
-    -webkit-mask: linear-gradient(#fff 0 0) content-box,
-      linear-gradient(#fff 0 0);
-    mask-composite: exclude;
-    pointer-events: none;
-  }
-
-  img {
-    max-width: 100%;
-    height: auto;
-  }
-
-  &.tokenAvatarNft {
-    border: none;
-    color: #fef08a;
-
-    &:before {
-      background: linear-gradient(
-        90deg,
-        rgba(240, 152, 25, 0.4),
-        rgba(237, 222, 93, 0.4)
-      );
-    }
-  }
-
-  &.tokenAvatarSft {
-    border: none;
-    color: #fed7aa;
-
-    &:before {
-      background: linear-gradient(
-        90deg,
-        rgba(240, 77, 25, 0.4),
-        rgba(237, 136, 93, 0.4)
-      );
-    }
-  }
-}
