+++ src/UI/SignTransactionsModals/SignWithDeviceModal/components/components/index.ts
@@ -1,0 +1,3 @@
+export * from './ConfirmAmount';
+export * from './ConfirmFee';
+export * from './ConfirmReceiver';
